chore: rename consts to styles, add index to page card component
