+++ packages/apps/human-app/frontend/src/modules/operator/views/profile/profile.page.tsx
@@ -4,8 +4,10 @@ import { useTranslation } from 'react-i18next';
 import { useIsMobile } from '@/shared/hooks/use-is-mobile';
 import { useGetKeys } from '@/modules/operator/hooks/use-get-keys';
 import { useWeb3AuthenticatedUser } from '@/modules/auth-web3/hooks/use-web3-authenticated-user';
-import { PageCardError } from '@/shared/components/ui/page-card/page-card-error';
-import { PageCardLoader } from '@/shared/components/ui/page-card/page-card-loader';
+import {
+  PageCardError,
+  PageCardLoader,
+} from '@/shared/components/ui/page-card';
 import { getErrorMessageForError } from '@/shared/errors';
 import { ProfileDisableButton } from '@/modules/operator/components/profile/profile-disable-button';
 import { ProfileListItem } from '@/shared/components/ui/profile-list-item';
