+++ packages/apps/human-app/frontend/src/modules/operator/views/sign-up/add-stake.page.tsx
@@ -2,9 +2,11 @@ import Grid from '@mui/material/Grid';
 import { Typography } from '@mui/material';
 import { useState } from 'react';
 import { t } from 'i18next';
-import { PageCardError } from '@/shared/components/ui/page-card/page-card-error';
-import { PageCardLoader } from '@/shared/components/ui/page-card/page-card-loader';
-import { PageCard } from '@/shared/components/ui/page-card/page-card';
+import {
+  PageCardError,
+  PageCardLoader,
+  PageCard,
+} from '@/shared/components/ui/page-card';
 import { getErrorMessageForError } from '@/shared/errors';
 import { Buttons } from '@/modules/operator/components/sign-up/add-stake/buttons';
 import { StakeForm } from '@/modules/operator/components/sign-up/add-stake/stake-form';
