+++ packages/apps/human-app/frontend/src/modules/worker/views/email-verification/email-verification.page.tsx
@@ -7,9 +7,11 @@ import { Button } from '@/shared/components/ui/button';
 import { routerPaths } from '@/router/router-paths';
 import { useVerifyEmailQuery } from '@/modules/worker/services/email-verification';
 import { SuccessLabel } from '@/shared/components/ui/success-label';
-import { PageCardError } from '@/shared/components/ui/page-card/page-card-error';
-import { PageCardLoader } from '@/shared/components/ui/page-card/page-card-loader';
-import { PageCard } from '@/shared/components/ui/page-card/page-card';
+import {
+  PageCardError,
+  PageCardLoader,
+  PageCard,
+} from '@/shared/components/ui/page-card';
 import { useLocationState } from '@/modules/worker/hooks/use-location-state';
 import { getErrorMessageForError } from '@/shared/errors';
