+++ packages/apps/human-app/frontend/src/modules/worker/views/hcaptcha-labeling/hcaptcha-labeling.page.tsx
@@ -9,8 +9,10 @@ import { env } from '@/shared/env';
 import { breakpoints } from '@/shared/styles/breakpoints';
 import { Counter } from '@/shared/components/ui/counter';
 import { useHCaptchaUserStats } from '@/modules/worker/services/hcaptcha-user-stats';
-import { PageCardError } from '@/shared/components/ui/page-card/page-card-error';
-import { PageCardLoader } from '@/shared/components/ui/page-card/page-card-loader';
+import {
+  PageCardError,
+  PageCardLoader,
+} from '@/shared/components/ui/page-card';
 import { getErrorMessageForError } from '@/shared/errors';
 import { useDailyHmtSpent } from '@/modules/worker/services/daily-hmt-spent';
 import { getTomorrowDate } from '@/shared/helpers/date';
