+++ packages/apps/human-app/frontend/src/modules/worker/views/sign/sign-up.page.tsx
@@ -15,7 +15,7 @@ import {
 import { Button } from '@/shared/components/ui/button';
 import { Input } from '@/shared/components/data-entry/input';
 import { Password } from '@/shared/components/data-entry/password/password';
-import { PageCard } from '@/shared/components/ui/page-card/page-card';
+import { PageCard } from '@/shared/components/ui/page-card';
 import { env } from '@/shared/env';
 import { getErrorMessageForError } from '@/shared/errors';
 import { Alert } from '@/shared/components/ui/alert';
