+++ packages/apps/human-app/frontend/src/shared/components/ui/page-card/index.ts
@@ -1,0 +1,3 @@
+export * from './page-card';
+export * from './page-card-error';
+export * from './page-card-loader';
